requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djoser==2.2.0
requests==2.31.0